Split "DevicePlugin" into its own file for enhanced modularity
